NarInfoCreate now fails on narSize == 0

+++ cachix-api/src/Cachix/Api/Signing.hs
@@ -19,21 +19,20 @@ import           Data.Text (Text)
 
 
 -- perl/lib/Nix/Manifest.pm:fingerprintPath
--- TODO: Either Text ByteString: assert values
 -- NB: references must be sorted
-fingerprint :: Text -> Text -> Int -> [Text] -> ByteString
+fingerprint :: Text -> Text -> Integer -> [Text] -> ByteString
 fingerprint storePath narHash narSize references = toS $ T.intercalate ";"
   ["1", storePath, narHash, T.pack (show narSize), T.intercalate "," references]
 
 -- Useful sinks for streaming nars
 
-sizeSink :: MonadIO m => Consumer ByteString m Int
-sizeSink = CC.foldM (\p n -> return (p + BS.length n)) 0
+sizeSink :: MonadIO m => Consumer ByteString m Integer
+sizeSink = CC.foldM (\p n -> return (p + fromIntegral (BS.length n))) 0
 
 hashSink :: MonadIO m => Consumer ByteString m (Context SHA256)
 hashSink = CC.foldM (\p n -> return (hashUpdate p n)) hashInit
 
-passthroughSizeSink :: MonadIO m => IORef Int -> Conduit ByteString m ByteString
+passthroughSizeSink :: MonadIO m => IORef Integer -> Conduit ByteString m ByteString
 passthroughSizeSink ioref = passthroughSink sizeSink (liftIO . writeIORef ioref)
 
 passthroughHashSink :: MonadIO m => IORef Text -> Conduit ByteString m ByteString
